Change mutex name to local scope in SimpleLogger
Using 'Global' prefix for mutex names requires SeCreateGlobalPrivilege which may not be available to all users. Updated the mutex name in the SimpleLogger class from a global scope to a local scope. This change restricts the mutex's accessibility to the local machine, enhancing the logging mechanism's isolation from other processes on different machines.