Add restrict-user job to build-and-release workflow
This commit introduces a new job named `restrict-user` in the `build-and-release.yml` workflow. The job runs on `ubuntu-latest` and includes a step that verifies if the GitHub actor is 'udlose'. If the actor does not match, a message is displayed, and the workflow exits with a status of 1, preventing unauthorized execution.