Update js-yaml.min.js as an embedded resource and provide a LogicalName to reference it from the embedded manifest resource stream
This commit adds `js-yaml.min.js` as an embedded resource in the project file with the logical name `MermaidPad.Assets.js-yaml.min.js`. It also removes the previous entries for `js-yaml.min.js` that were listed as a `None` type and as an `EmbeddedResource`, indicating that the file is no longer treated as a resource in the project. #38