Refactor GetSvgDimensionsAsync to use ReadOnlyMemory<char>
Updated the IImageConversionService interface to change the
GetSvgDimensionsAsync method parameter from `string` to
`ReadOnlyMemory<char>` for improved memory efficiency.

Removed the previous implementation of GetSvgDimensionsAsync
in SkiaSharpImageConversionService and replaced it with a
`NotImplementedException` and placeholder code. Added `TODO`
comments to highlight the need for a more efficient approach
to determine SVG dimensions without fully loading the SVG.